+++ invoice/src/main/java/az/cybernet/invoice/dto/response/FilteredInvoiceResp.java
@@ -5,6 +5,7 @@ import az.cybernet.invoice.enums.Status;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -18,7 +19,7 @@ public class FilteredInvoiceResp {
     UUID senderId;
     UUID customerId;
     Status status;
-    Double total;
+    BigDecimal total;
     LocalDateTime createdAt;
     LocalDateTime updatedAt;
     String comment;
