+++ invoice/src/main/java/az/cybernet/invoice/dto/response/InvoiceDetailResponse.java
@@ -4,6 +4,7 @@ import az.cybernet.invoice.enums.Status;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -20,7 +21,7 @@ public class InvoiceDetailResponse {
     UUID senderId;
     UUID customerId;
     Status status;
-    Double total;
+    BigDecimal total;
     LocalDateTime createdAt;
     LocalDateTime updatedAt;
     String comment;
